fix: corrupt yml file
In path themes/next/languages existed a file called default.yml, which
used to be a symbolic link. Somehow, the file corrupted for some reason.

It looks normal in Windows systems but not in Ubuntu Linux. Cause the
hexo couldn't stat the file normally in the linux environment then
terminated the generate progress. GitHub Actions failed because of it.